Use specific int types (stdint), add problem 3

+++ include/solutions/p001.h
@@ -1,6 +1,8 @@
 #ifndef P001_H
 #define P001_H
 
-long solve_p1(void);
+#include <stdint.h>
+
+uint32_t solve_p1(void);
 
 #endif
